# contracts/api/openapi.yaml
...
components:
  schemas:
...
    Body_upload_document_api_jobs_documents_post:
      properties:
        file:
          format: binary
          title: File
          type: string
        published_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Published At
        security_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Security Id
        thesis_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Thesis Id
        view:
          default: ''
          title: View
          type: string
      required:
      - file
      title: Body_upload_document_api_jobs_documents_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    JobAcceptedOut:
      properties:
        document_id:
          title: Document Id
          type: string
        job_id:
          title: Job Id
          type: string
        status:
          default: queued
          title: Status
          type: string
      required:
      - job_id
      - document_id
      title: JobAcceptedOut
      type: object
    JobStatusOut:
      properties:
        enqueue_time:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Enqueue Time
        finish_time:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finish Time
        job_id:
          title: Job Id
          type: string
        result:
          title: Result
        start_time:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Start Time
        status:
          title: Status
          type: string
        success:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Success
      required:
      - job_id
      - status
      title: JobStatusOut
      type: object
...
    ProcessingJobOut:
      properties:
        attempt_count:
          title: Attempt Count
          type: integer
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        document_id:
          title: Document Id
          type: string
        finished_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finished At
        job_id:
          title: Job Id
          type: string
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        max_attempts:
          title: Max Attempts
          type: integer
        result:
          anyOf:
          - type: object
          - type: 'null'
          title: Result
        security_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Security Id
        source_filename:
          title: Source Filename
          type: string
        started_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Started At
        status:
          title: Status
          type: string
      required:
      - job_id
      - document_id
      - source_filename
      - security_id
      - status
      - attempt_count
      - max_attempts
      - result
      - last_error
      - created_at
      - started_at
      - finished_at
      title: ProcessingJobOut
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: 主动权益投资逻辑智能协作平台。系统输出候选信号与状态建议，不产生任何交易、评级或调仓指令。
  title: AI Investment Copilot
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/jobs:
    get:
      operationId: list_jobs_api_jobs_get
      parameters:
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProcessingJobOut'
                title: Response List Jobs Api Jobs Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Jobs
      tags:
      - jobs
  /api/jobs/documents:
    post:
      operationId: upload_document_api_jobs_documents_post
      parameters:
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_jobs_documents_post'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Upload Document
      tags:
      - jobs
  /api/jobs/{job_id}:
    get:
      operationId: get_job_api_jobs__job_id__get
      parameters:
      - in: path
        name: job_id
        required: true
        schema:
          title: Job Id
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Job
      tags:
      - jobs
  /api/jobs/{job_id}/replay:
    post:
      operationId: replay_job_api_jobs__job_id__replay_post
      parameters:
      - in: path
        name: job_id
        required: true
        schema:
          title: Job Id
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Replay Job
      tags:
      - jobs
...